+++ backend/controller/uploadController.ts
@@ -3,7 +3,8 @@ import { Request as ExRequest, Response as ExResponse, NextFunction } from 'expr
 import fs from 'node:fs/promises'
 import { Body, Consumes, Controller, Get, Middlewares, Post, Produces, Query, Request, Route, SuccessResponse } from 'tsoa'
 import { _id } from '../../common/types.js'
-import { documentFileHandler, fileHandler, getSettings } from '../helper.js'
+import { getSettings } from '../db.js'
+import { documentFileHandler, fileHandler } from '../helper.js'
 import i18n from '../i18n.js'
 import Token from '../models/token.js'
 import User from '../models/user.js'
