+++ backend/models/healthCareCost.ts
@@ -7,7 +7,8 @@ import {
   baseCurrency,
   healthCareCostStates
 } from '../../common/types.js'
-import { convertCurrency, costObject } from './helper.js'
+import { convertCurrency } from './exchangeRate.js'
+import { costObject } from './helper.js'
 import { ProjectDoc } from './project.js'
 
 interface Methods {
